display: Define soong namespace
HALs are starting to move to blueprint files, which means that pathmap
ifeqs no longer work. This results in modules with the same name in
different platform HAL directories. Since these modules aren't used or
linked anywhere outside the HAL directories and a device's device.mk,
we can safely use a soong namespace to isolate the different HAL
directories' blueprints.

Change-Id: I86ebbf1e04a379d641444d74b90df0368f90f659
Signed-off-by: Yash-Garg <ben10.yashgarg@gmail.com>